// third-party/ketcher/ketcher-react/src/ui/toolbars/toolbar.types.ts
type TopToolbarItemVariant =
  | 'clear'
  | 'open'
  | 'save'
  | 'undo'
  | 'redo'
  | 'cut'
  | 'copies'
  | 'copy'
  | 'copy-image'
  | 'copy-mol'
  | 'copy-ket'
  | 'paste'
  | 'zoom-in'
  | 'zoom-out'
  | 'zoom-list'
  | 'layout'
  | 'clean'
  | 'arom'
  | 'dearom'
  | 'cip'
  | 'check'
  | 'analyse'
  | 'calculation-readiness'
  | 'recognize'
  | 'miew'
  | 'settings'
  | 'help'
  | 'about';

type LeftToolbarItemVariant =
  // select group
  | 'hand'
  | 'select'
  | 'select-lasso'
  | 'select-rectangle'
  | 'select-structure'
  | 'select-fragment'
  | 'erase'
  // bond group
  | 'bonds'
  | 'bond-common'
  | 'bond-single'
  | 'bond-double'
  | 'bond-triple'
  | 'bond-dative'
  | 'bond-stereo'
  | 'bond-up'
  | 'bond-down'
  | 'bond-updown'
  | 'bond-crossed'
  | 'bond-query'
  | 'bond-special'
  | 'bond-any'
  | 'bond-hydrogen'
  | 'bond-aromatic'
  | 'bond-singledouble'
  | 'bond-singlearomatic'
  | 'bond-doublearomatic'
  | 'chain'
  // charge group
  | 'charge-plus'
  | 'charge-minus'
  // sgroup group
  | 'sgroup'
  // reaction
  // plus
  | 'reaction-plus'
  // arrows
  | 'arrows'
  | 'reaction-arrow-open-angle'
  | 'reaction-arrow-filled-triangle'
  | 'reaction-arrow-filled-bow'
  | 'reaction-arrow-dashed-open-angle'
  | 'reaction-arrow-failed'
  | 'reaction-arrow-retrosynthetic'
  | 'reaction-arrow-both-ends-filled-triangle'
  | 'reaction-arrow-equilibrium-filled-half-bow'
  | 'reaction-arrow-equilibrium-filled-triangle'
  | 'reaction-arrow-equilibrium-open-angle'
  | 'reaction-arrow-unbalanced-equilibrium-filled-half-bow'
  | 'reaction-arrow-unbalanced-equilibrium-open-half-angle'
  | 'reaction-arrow-unbalanced-equilibrium-large-filled-half-bow'
  | 'reaction-arrow-unbalanced-equilibrium-filled-half-triangle'
  | 'reaction-arrow-elliptical-arc-arrow-filled-bow'
  | 'reaction-arrow-elliptical-arc-arrow-filled-triangle'
  | 'reaction-arrow-elliptical-arc-arrow-open-angle'
  | 'reaction-arrow-elliptical-arc-arrow-open-half-angle'
  | typeof MULTITAIL_ARROW_TOOL_NAME
  // mapping
  | 'reaction-mapping-tools'
  | 'reaction-automap'
  | 'reaction-map'
  | 'reaction-unmap'
  // rgroup group
  | 'rgroup'
  | 'rgroup-label'
  | 'rgroup-fragment'
  | 'rgroup-attpoints'
  // create monomer
  | 'create-monomer'
  // shape group
  | 'shapes'
  | 'shape-ellipse'
  | 'shape-rectangle'
  | 'shape-line'
  // text group
  | 'text'
  // image group
  | typeof IMAGE_KEY;

type BottomToolbarItemVariant =
  | 'template-common'
  | 'template-lib'
  | 'enhanced-stereo'
  | 'fullscreen';

type RightToolbarItemVariant =
  | 'atom'
  | 'freq-atoms'
  | 'period-table'
  | 'extended-table'
  | 'any-atom';

type FloatingToolItemVariant =
  | 'transform-flip-h'
  | 'transform-flip-v'
  | 'erase';

type ToolbarItemVariant =
  | TopToolbarItemVariant
  | LeftToolbarItemVariant
  | BottomToolbarItemVariant
  | RightToolbarItemVariant
  | FloatingToolItemVariant
  | 'bonds';

interface ToolbarItem {
  id: ToolbarItemVariant;
  options?: ToolbarItem[];
}